+++ src/actions/create-action.ts
@@ -1,9 +1,9 @@
 import { ActionConstant } from '../types/actions';
 import { CreateAction } from '../types/create-action';
-import { isFailure, getValue } from './failure';
 import createAsyncAction from './create-async-action';
 import bindActionType from './bind-action-type';
 import validateActionType from './validate-action-type';
+import { isKnownError } from '../utils/errors';
 
 // Synchronous action creator.
 //
@@ -17,7 +17,7 @@ const createAction = (
 ) => {
   validateActionType(actionType);
 
-  const executeEffectAndReturnAction = (input: any) => {
+  function executeEffectAndReturnAction(input: any) {
     if (effect === undefined) {
       if (input === undefined) {
         return { type: actionType };
@@ -26,20 +26,27 @@ const createAction = (
       return { type: actionType, payload: input };
     }
 
-    const payload = effect(input);
+    return { type: actionType, payload: effect(input) };
+  }
 
-    if (isFailure(payload)) {
-      return {
-        type: actionType,
-        error: true,
-        payload: getValue(payload),
-      };
+  return bindActionType(actionType, function* (input: any) {
+    try {
+      const action = executeEffectAndReturnAction(input);
+      yield action;
+      return action;
+    } catch (error) {
+      const action = { type: actionType, error: true, payload: error };
+      yield action;
+
+      // Known errors are swallowed to avoid false positives in the console
+      // and error reporting services. This behavior is opt-in only.
+      if (isKnownError(error)) {
+        return action;
+      } else {
+        throw error;
+      }
     }
-
-    return { type: actionType, payload };
-  };
-
-  return bindActionType(actionType, executeEffectAndReturnAction);
+  });
 };
 
 // Asynchronous action creator.